lift and shift csv views to reports from api
