+++ tock/api/tests.py
@@ -24,7 +24,6 @@ from django.test.client import Client
 from rest_framework.test import APIClient
 
 # common client for all API tests
-
 def client(self):
     self.request_user = User.objects.get_or_create(username='aaron.snow')[0]
     self.token = Token.objects.get_or_create(user=self.request_user)[0].key
@@ -39,16 +38,11 @@ FIXTURES = [
     'hours/fixtures/timecards.json'
 ]
 
-
 class ProjectsAPITests(TestCase):
     fixtures = FIXTURES
 
     def test_projects_json(self):
         pass
-
-    def test_projects_csv(self):
-        pass
-
 
 class ProjectInstanceAPITests(WebTest):
     fixtures = FIXTURES
@@ -68,22 +62,14 @@ class UsersAPITests(TestCase):
     def test_users_csv(self):
         pass
 
-
 class TimecardsAPITests(WebTest):
     fixtures = FIXTURES
 
     def test_timecards_json(self):
         """ Check that the timecards are rendered in json format correctly """
-        res = client(self).get(reverse('TimecardList', kwargs={'format': 'json'})).content
+        res = client(self).get(reverse('TimecardList')).content
         clean_res = json.loads(res.decode())
         self.assertEqual(clean_res['count'], 2)
-
-    def test_timecards_csv(self):
-        """ Check that the timecards are rendered in csv format correctly """
-        res = client(self).get(reverse('TimecardList', kwargs={'format': 'csv'})).content
-        target_fields = len(TimecardSerializer.__dict__['_declared_fields'])
-        output_fields = len(str(res).split('\\r')[0].split(','))
-        self.assertEqual(target_fields, output_fields)
 
     # TODO: test with more diverse data
     def test_get_timecards(self):
@@ -138,72 +124,6 @@ class TimecardsAPITests(WebTest):
             params={'submitted': 'foo'}
         )
         self.assertEqual(len(queryset), 2)
-
-
-class ProjectTimelineTests(WebTest):
-    fixtures = FIXTURES
-
-    def test_project_timeline(self):
-        res = client(self).get(reverse('UserTimelineView'))
-        self.assertIn(
-            'aaron.snow,2015-06-01,2015-06-08,False,20.00', str(res.content))
-
-
-class BulkTimecardsTests(TestCase):
-    fixtures = FIXTURES
-
-    def test_bulk_timecards(self):
-        response = client(self).get(reverse('BulkTimecardList'))
-        rows = decode_streaming_csv(response)
-        expected_fields = set((
-            'project_name',
-            'project_id',
-            'billable',
-            'employee',
-            'start_date',
-            'end_date',
-            'hours_spent',
-            'agency',
-            'flat_rate',
-            'active',
-            'mbnumber',
-            'notes',
-        ))
-        rows_read = 0
-        for row in rows:
-            self.assertEqual(set(row.keys()), expected_fields)
-            self.assertEqual(row['project_id'], '1')
-            rows_read += 1
-        self.assertNotEqual(rows_read, 0, 'no rows read, expecting 1 or more')
-
-class BulkTimecardsTests(TestCase):
-    fixtures = FIXTURES
-
-    def test_slim_bulk_timecards(self):
-        response = client(self).get(reverse('SlimBulkTimecardList'))
-        rows = decode_streaming_csv(response)
-        expected_fields = set((
-            'project_name',
-            'billable',
-            'employee',
-            'start_date',
-            'end_date',
-            'hours_spent',
-            'mbnumber',
-        ))
-        rows_read = 0
-        for row in rows:
-            self.assertEqual(set(row.keys()), expected_fields)
-            self.assertEqual(row['project_name'], 'Out Of Office')
-            self.assertEqual(row['billable'], 'False')
-            rows_read += 1
-        self.assertNotEqual(rows_read, 0, 'no rows read, expecting 1 or more')
-
-
-def decode_streaming_csv(response, **reader_options):
-    lines = [line.decode('utf-8') for line in response.streaming_content]
-    return csv.DictReader(lines, **reader_options)
-
 
 class TestAggregates(WebTest):
 
@@ -273,9 +193,7 @@ class TestAggregates(WebTest):
 
     def test_hours_by_quarter_by_user_with_unsubmitted_timecards(self):
         """ Check that unsubmitted timecards are not counted  """
-
         # add one unsubmitted timecard + one additional submitted one
-
         timecard_unsubmit = TimecardFactory(
             user=self.user,
             reporting_period=ReportingPeriodFactory(
@@ -311,7 +229,6 @@ class TestAggregates(WebTest):
 
         self.assertEqual(len(self.timecard_objects), 4)
         self.assertEqual(row['total'], 60)
-
 
 class ReportingPeriodList(WebTest):
     fixtures = FIXTURES
@@ -350,7 +267,6 @@ class ReportingPeriodList(WebTest):
             )
         ).data
         self.assertEqual(res['count'], 1)
-
 
     def test_ReportingPeriodList_json_no_longer_employed(self):
         """ Check that the ReportingPeriodList shows users that have missing
