+++ tock/hours/urls/reports.py
@@ -1,11 +1,11 @@
 from django.conf.urls import url
 
 from .. import views
-from api.views import bulk_timecard_list, slim_bulk_timecard_list
+from api.views import ProjectList
 
 urlpatterns = [
-    url(r'^timecards_bulk.csv$', bulk_timecard_list, name='BulkTimecardList'),
-    url(r'^slim_timecard_bulk.csv$', slim_bulk_timecard_list, name='SlimBulkTimecardList'),
+    url(r'^timecards_bulk.csv$', views.bulk_timecard_list, name='BulkTimecardList'),
+    url(r'^slim_timecard_bulk.csv$', views.slim_bulk_timecard_list, name='SlimBulkTimecardList'),
     url(regex=r'^$', view=views.ReportsList.as_view(), name='ListReports'),
     url(regex=r'^(?P<reporting_period>[0-9]{4}-[0-9]{2}-[0-9]{2})/$',
         view=views.ReportingPeriodDetailView.as_view(), name='ReportingPeriodDetailView'),
@@ -13,4 +13,8 @@ urlpatterns = [
         view=views.ReportingPeriodCSVView, name='ReportingPeriodCSVView'),
     url(regex=r'^(?P<reporting_period>[0-9]{4}-[0-9]{2}-[0-9]{2})/(?P<username>[A-Za-z0-9._%+-]*)/$',
         view=views.ReportingPeriodUserDetailView.as_view(), name='ReportingPeriodUserDetailView'),
+    url(r'^project_timeline.csv$', views.project_timeline_view, name='ProjectTimelineView'),
+    url(r'^user_timeline.csv$', views.user_timeline_view, name='UserTimelineView'),
+#    url(r'^projects.csv$', ProjectList.as_view(), name='ProjectList'),
+
 ]
